backend/requirements.txt
autopep8==2.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
PyJWT==2.3.0
python-dotenv==0.21.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2